file replacement

+++ test/FamilyDAO.t.sol
@@ -34,11 +34,17 @@ contract FamilyDAOTest is Test {
         registry.setFakeMember(daughter, true, "daughter", admin);
         registry.setFakeMember(son, true, "son", admin);
 
-        nft.mint(admin, "https://pink-labour-ox-753.mypinata.cloud/ipfs/bafkreichiugv37vbqdrbhlizcpldcfiquwkfpjxslbqrv2jzxjyai2dzia?pinataGatewayToken=m-OsOwJyM0daP06Gyqh9UDfa9__UgsIT22Z0br4IZ2h-LrIvY5w5H3FWN8E15G3J");
+        nft.mint(
+            admin,
+            "https://pink-labour-ox-753.mypinata.cloud/ipfs/bafkreichiugv37vbqdrbhlizcpldcfiquwkfpjxslbqrv2jzxjyai2dzia?pinataGatewayToken=m-OsOwJyM0daP06Gyqh9UDfa9__UgsIT22Z0br4IZ2h-LrIvY5w5H3FWN8E15G3J"
+        );
 
         string memory uri = nft.tokenURI(0);
         emit log_string(uri);
-        assertEq(uri, "https://pink-labour-ox-753.mypinata.cloud/ipfs/bafkreichiugv37vbqdrbhlizcpldcfiquwkfpjxslbqrv2jzxjyai2dzia?pinataGatewayToken=m-OsOwJyM0daP06Gyqh9UDfa9__UgsIT22Z0br4IZ2h-LrIvY5w5H3FWN8E15G3J");
+        assertEq(
+            uri,
+            "https://pink-labour-ox-753.mypinata.cloud/ipfs/bafkreichiugv37vbqdrbhlizcpldcfiquwkfpjxslbqrv2jzxjyai2dzia?pinataGatewayToken=m-OsOwJyM0daP06Gyqh9UDfa9__UgsIT22Z0br4IZ2h-LrIvY5w5H3FWN8E15G3J"
+        );
 
         nft.mint(daughter, "ipfs://daughter-nft");
         nft.mint(son, "ipfs://son-nft");
@@ -110,16 +116,10 @@ contract FamilyDAOTest is Test {
 
         vm.startPrank(admin);
 
-        string memory desc = string.concat(
-            "ADD_MEMBER:",
-            vm.toString(cousin),
-            ":",
-            vm.toString(admin),
-            ":cousin:",
-            "ipfs://dummy-uri"
-        );
+        string memory desc =
+            string.concat("ADD_MEMBER:", vm.toString(cousin), ":", vm.toString(admin), ":cousin:", "ipfs://dummy-uri");
 
-        uint proposalId = dao.createProposal(desc);
+        uint256 proposalId = dao.createProposal(desc);
 
         vm.stopPrank();
 
@@ -150,7 +150,7 @@ contract FamilyDAOTest is Test {
 
         string memory desc = string.concat("REMOVE_MEMBER:", vm.toString(cousin));
         vm.prank(admin);
-        uint id = dao.createProposal(desc);
+        uint256 id = dao.createProposal(desc);
 
         vm.prank(daughter);
         dao.vote(id, true);
